toml: combine tests back into one file
