+++ src/plugins/toml/CMakeLists.txt
@@ -146,24 +146,8 @@ if (DEPENDENCY_PHASE)
 endif (DEPENDENCY_PHASE)
 
 add_plugin (
-	toml INSTALL_TEST_DATA TEST_README
+	toml ADD_TEST INSTALL_TEST_DATA TEST_README
 	TEST_REQUIRED_PLUGINS type base64
 	LINK_ELEKTRA elektra-meta
 	SOURCES ${SOURCE_FILES}
 	INCLUDE_DIRECTORIES ${CMAKE_CURRENT_SOURCE_DIR} COMPONENT libelektra${SO_VERSION})
-
-if (ADDTESTING_PHASE)
-	add_plugintest (
-		toml
-		LINK_PLUGIN toml
-		LINK_ELEKTRA elektra-meta
-		INCLUDE_DIRECTORIES ${CMAKE_CURRENT_SOURCE_DIR}
-		INSTALL_TEST_DATA)
-
-	add_plugintest (
-		toml_root
-		LINK_PLUGIN toml
-		LINK_ELEKTRA elektra-meta
-		INCLUDE_DIRECTORIES ${CMAKE_CURRENT_SOURCE_DIR}
-		INSTALL_TEST_DATA)
-endif ()
