--- src/plugins/toml/testmod_toml_root.c
@@ -1,12 +1,0 @@
-/**
- * @file
- *
- * @brief Tests for the toml plugin
- *
- * @copyright BSD License (see LICENSE.md or https://www.libelektra.org)
- *
- */
-
-#define PREFIX "user:/"
-
-#include "testmod_toml.inc.c"
